restracture project, changes imports, added styles files
