+++ src/App.jsx
@@ -1,5 +1,5 @@
-import { Layout } from './layout/Layout';
-import { HomePage } from './pages/HomePage';
+import { Layout } from '@layout/Layout/Layout';
+import { HomePage } from '@pages/HomePage/HomePage';
 
 export const App = () => {
   return (
