--- src/layout/Footer.jsx
@@ -1,3 +1,0 @@
-export const Footer = () => {
-  return <footer>@2025</footer>;
-};
