+++ src/layout/Footer/Footer.jsx
@@ -1,0 +1,9 @@
+import styles from './Footer.module.css';
+
+export const Footer = () => {
+  return (
+    <footer className={styles.footer}>
+      @2025
+    </footer>
+  );
+};
