--- src/layout/Header.jsx
@@ -1,3 +1,0 @@
-export const Header = () => {
-  return <header>Pick restaurant and order food</header>;
-};
