+++ src/layout/Header/Header.jsx
@@ -1,0 +1,9 @@
+import styles from './Header.module.css';
+
+export const Header = () => {
+  return (
+    <header className={styles.header}>
+      Pick restaurant and order food
+    </header>
+  );
+};
