+++ src/layout/Header/Header.module.css
@@ -1,0 +1,5 @@
+.header {
+	grid-area: header;
+	padding: 12px;
+	border: 1px solid black;
+}
