+++ src/layout/Layout/Layout.module.css
@@ -1,0 +1,10 @@
+.layout {
+  display: grid;
+  grid-template-rows: auto 1fr auto;
+  grid-template-areas:
+    'header'
+    'main'
+    'footer';
+  height: 100vh;
+  margin: 0;
+}
