+++ src/layout/TabNavigation/TabNavigation.module.css
@@ -1,0 +1,11 @@
+.container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 12px;
+  padding: 24px;
+}
+
+.selected {
+  border-bottom: 1px solid black;
+}
